backend/requirements.txt
Flask==3.0.3
pymongo==4.10.1
cryptography==43.0.3
python-dotenv==1.0.1
pytz

